http: use output stream to return files
Reading an entire file can consume a lot of memory, this patch change
the way files are streamed in the http server.

Instead of reading an entire file and assign it to the reply string, the
write_body with stream option is used.

The file_handler support transformers that allows users of the handler
to modify a file, currently there is only one such transformer that does
string replace.

This patch also change the way transformer behave so that existing code
will continue to work.

file_transformer will now act as a factory class for output_stream.

Different transformer will inherit the file_transformer and would wrap
the output stream with a new stream that would do the transformation.

The implementation of content_replace was modified so that users of that
class would get the same result with the new file_handler behavior.

To efficiently do a transformation, a helper class that acts as a simple
final state machine was added: buffer_replace, it was optimized for
temporary_buffer and the {{key}} format.

buffer_replace holds a set of key/value to replace.
Its replace method gets a temporary_buffer, search for the opening
brace and if it finds it, it would follow the keys it has and stop
when a comparison fails.

It does not take additional space when searching for the keys and does
not holds the keys in advance as a parser would do.

tests where intentionally removed and will be added on the next patch

See scylladb/scylla#2816

Signed-off-by: Amnon Heiman <amnon@scylladb.com>