iostream: Add a helper method to copy one stream to another
This patch adds a copy method that gets an input stream and an output
stream, it consumes all the input stream and pass its content to the
output stream.

A helper struct stream_copy_consumer is used for the consume method.

For compilation time optimization, the templates is initilized in
fstream.cc

Signed-off-by: Amnon Heiman <amnon@scylladb.com>