execution_stage: defer execution less aggressively
Instead of accumulating 128 tasks before flushing, flush
immediately.

Since a flush doesn't run immediately, and only schedules
execution of the batch for later, tasks that are already
queued will have the opportunity to increase the batch size.

On the other hand, with the current behavior, batches that are
smaller than 128 are deferred to then next poll, which can be
a task_quota in the future. Since some operations require many
stages, we end up increasing latency by nr_stages * task_quota,
in the presence of a CPU hog. With early flush, latency is kept
at task_quota.
Message-Id: <20171017164611.7711-1-avi@scylladb.com>