test httpd: Add test for the transformer
This patch replaces the http transformer test to the new output stream
format.

It adds a sink helper class that can make an output stream collect
evertyhing allowing to verify the output.

The test, check that replacement is done only on a specific file
extenstion, and different edge cases.

A helper function was added to load the trnsformer with buffer
fragments.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>