Adding dpdk-port-index as a command line option with default value of 0
DPDK in order to use certain network interface has to be unbound from the kernel and bound to DPDK itself.
DPDK can have mulitple interfaces bound, therefore there is a need for an option to select the particular one.
Possible sceanario would be a vanilla DPDK application and SEASTAR app running in paraller.
THe Port index is the index (starting from 0) of interfaces bound to DPDK shown by dpdk-setup.sh

Message-Id: <1509920787-4431-1-git-send-email-marek.waszkiewicz77@gmail.com>