tutorial: a section about when_all() and when_all_succeed()
This adds a section to the tutorial introducing the variadic variants
of when_all() and when_all_succeed(). Sadly, the verboseness and, may I say,
ugliness, of the first function is very clear here. But let's at least
describe what we have right now (and possibly forever, given that this function
is based on a standard C++ one).

The vector variants of these functions are not yet explained.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20171113152311.7166-1-nyh@scylladb.com>