provide an interface for updating the shares of an I/O class.
The Fair Queue provides an API that allows the share of a class to be
runtime-updated, but to date the I/O queue has not taken advantage of
that.

This patch does that by allowing any caller to update the shares of a
class. The shares are updated for the calling shard only.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
---
v2:
- returns a future so it doesn't go dangling
- do not use a reference.