http: api_docs add swagger 2.0 support
The main difference between swagger 1.2 and swagger 2.0 is the use of a
single file for the swagger definition.

A single file makes the API hard to maintain and makes it harder to
enable a bunch of APIs together, a feature that is useful during
startup.

In Swagger 1.2, the api_docs class generates the swagger file for the
API table of content. The Swagger-UI and other clients read the table of
content file and use it to fetch the rest of the API.

This patch adds a swagger 2.0 support that mimic this behavior.

In the swagger 2.0 all the apis are under the paths and all the type
definitions are under "definitions".

This patch create a parallel class hierarchy for swagger 2.0 were
mutual functionality was put in a base classes.

api_registry_20 is a handler class that generates a unified file, it
uses api_docs_20 to holds the different doc entry.

The registration in 2.0 is done with a doc_entry which is a
noncopyable_function. This also adds dynamic API support, allowing
swagger definition generation dynamically.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <20171207064743.19573-1-amnon@scylladb.com>