core: Introduce cancellation mechanism through class abort_source
The abort_source class enables one to communicate a cancellation
request to an ongoing asynchronous operations.

Such an operation can subscribe to the cancellation request by
registering a callback that will be invoked when
abort_source::request_abort() is called.

This mechanism implement interruptible actions, such as an
abortable sleep.

Signed-off-by: Duarte Nunes <duarte@scylladb.com>