reactor: get rid of reactor_backend::abort_reader/abort_writer
abort_reader()/abort_writer() cause epoll to stop waiting for the
respective events even though the initiating non-blocking system
call was not undone. It is used for UDP reads and writes, and
for accept().

However, it turns out that we can undo UDP reads and writes
as well as accept()s by calling shutdown() on the socket. The
shutdown() call returns with an error, but it does cause the next
epoll_wait() to return with that fd in the epoll event list.

This patch gets rid of abort_read() and abort_write() by calling
shutdown() instead. The operations that were waited on that fd
becoming readable or writeable will be woken up and retried.

A complication is that after shutdown() on a listening socket,
epoll_wait() return with EPOLLOUT|EPOLLHUP, even though during normal
operation (when accept() would have succeeded) it returns POLLIN. To
handle this inconsistency we introduce readable_or_writeable(), which
returns a future that becomes ready when the fd is either readable or
writeable, to complement the existing readable() and writeable()
functions.

Another complication is that a pending recv() or send() will return
with EAGAIN, which normally just says that the buffer space was
exhausted. To distinguish betweem the two cases, add a flag and check
it.

The motivation for this patch is an alternative poll implementation
using io_submit()/io_getevents() rather than epoll. While there is
an equivalent to abort_read() (io_cancel()), it is not implemented
efficiently, and is asynchronous, so a change here is needed anyway.