future-util: optimise parallel_for_each()
It is not uncommon that in CPU-bound workloads most of the futures are
ready. In such cases it is especially important that utilities like
parallel_for_each() are sufficiently optimised.

This patch optimistically optimises for situations when all futures
processed by parallel_for_each() are immediately ready. The penalty
when that is not true is expected to be negligible.

Message-Id: <20180105190107.8926-1-pdziepak@scylladb.com>