rwlock: add method for determining if an rwlock is locked
This patch adds a method rwlock::locked() for determining whether an rwlock
is currently locked (for either read or write). This is useful when, for
example, code wants to delete a lock once nobody is waiting on it.

Since semaphore also has a similar method for querying its state, it only
makes sense that rwlock should too. In fact, rwlock::lock() just uses
semaphore::available_units().

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180108161737.12037-1-nyh@scylladb.com>