prometheus: Support metric family name parameter
This patch adds an ability to return specific metrics family from
prometheus.

When using a browser or curl for the text versions, the following would
work:

http://localhost:9180/metrics - return all metrics
http://localhost:9180/metrics/cache* - return all metrics that starts
with cache (metrics names are passed without the prefix)
http://localhost:9180/metrics/cache_active_reads return a specific
metric

To set the prometheus server to read metrics, change the prometheus.yml
file, as described in doc/prometheus.md.

The implementation change the metric_family_range to accept iterators.
And make metrics_families_per_shard a class with lower_bound and
upper_bound method to return the first and end iterator of
a range.

The get_range helper function, create a range based on the metric name
and if a prefix needs to be used.

The implementation of metric_family_iterator was modified to act more
'iterator' like. A minor optimization was removed in favor of the ++
operator to act like expected from an iterator.

A helper method, check for the asterisk at the end of the metric name.
Because Prometheus server would encode-url the path, the search is both
for the '*' symbol and for the encoded representation: '%2A'.
Message-Id: <20180111101751.9899-2-amnon@scylladb.com>