future-util: failure to allocate internal state is unrecoverable
parallel_for_each, repeat, repeat_until_value and do_until require an
internal state to handle cases when the execution is suspended. However,
they are all optimised for cases when that doesn't happen and,
consequently, the allocation is done lazily.

Unfortunately, the current implementation make them exception unsafe. If
the allocation of the internal state fails they return while leaving an
unwaited continuation.

The solution that doesn't involve any pessimisation and follows the
approach taken by future<...>::then() is to assume that such failures
are not possible to recover from.

Fixes #384.

Message-Id: <20180111161904.16984-1-pdziepak@scylladb.com>