Prometheus should check that the iterator is valid before using it
When searching for a specific metric family name, the returned iterator could
be invalid (if the searched metric family name is lexically bigger than all
metircs family names).

We should check that the return iterator is valid before trying to check
that it points to the right metric family name.

Fixes #387

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <20180114152021.29800-1-amnon@scylladb.com>