systemwide_memory_barrier: don't use mprotect() on ARM
Some ARM processors support a TLBI instruction variant that can broadcast
a TLB invalidate to all cores. On those processors, mprotect() will not
generate inter-processor interrupts (IPIs), and therefore memory barriers
will not be injected to other reactors, defeating the purpose of
systemwide_memory_barrier().

To fix, fail if we were not able to use sys_membarrier(). Return false
indicating that we failed; the reactor will simply poll forever until
an event is available or the kernel is upgraded.
Message-Id: <20180121145602.10746-1-avi@scylladb.com>