rwlock: add exception-safe lock/unlock alternative
semaphore has, in addition to the separate wait/signal methods,
a get_units() function which returns an owning object, so when this object
is destroyed for any reason, the operation is undone. This makes it easier
to write exception-safe code.

rwlock only has separate lock/unlock methods, and until now didn't have
a get_units()-like mechanism. This patch adds it.

Because rwlock is implemented using a semaphore, the implemenation is
trivial, and uses the existing semaphore's get_units and related types.
We also add an alias, rwlock::holder, for the returned object type.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180125131339.32151-1-nyh@scylladb.com>