tutorial: note about asynchronous functions throwing exceptions
In the "Exceptions vs. exceptional futures" I recommended that
asynchronous functions (functions returning a future) should not
throw, and rather should return an exceptional future.
In this patch, I add a note that this is a recommendation, but
cannot be assumed because even Seastar methods like semaphore::wait()
violate this recommendation.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180131204930.21850-1-nyh@scylladb.com>