reactor: change adjustment method for tasks becoming active
The CPU scheduler has the role of adjusting for how long task groups run
over time according to the priority (we call them shares) the task
groups have.

To make sure task groups with different shares are comparable, we keep
track not only of their runtime but also their virtual runtime
(vruntime), which is essentially the runtime adjusted to the amount of
shares that task group has over the total.

If all task groups are constantly running, all we have to do is
guarantee that we always pick the task group with the lower vruntime to
run next. But in reality, not all task groups run all the time and they
may spend some time idle. One easy way to understand that is a bursty
workload of reads or writes in a database. Another example are
background processes like compactions and repairs in Scylla: at any
given time they may or may not have work to do.

When a task group doesn't run, be it for a long time or just a couple
of task_quotas, its vruntime will become a lot lower than the others and
it will dominate the CPU for a long time to come. To prevent that, we
forget the existing vruntime when a task is made active, and
artificially set it to "the last vruntime a task group had, minus a task
quota".

This is problematic for two reasons: first, we are assuming that tasks
will respect the task quota, or at least be close to that. In reality,
we are painfully aware that this is not true. Second, a task group with
a low share amount can accrue a lot of vruntime in one run, and if the
task group waking up is now set to have that vruntime, other task groups
that weren't inactive may run for a long time.

Let's look at a real example of this happening in a Scylla run with a
250usec task quota for a write only workload where we want to
CPU-isolate writes (1000 shares), memtable flushes (~10 shares) and
compactions (100 shares).

I could observe in practice the following sequence of events:

- memtable flushes are a backlog-controlled process, and their shares
  are quite small.
- writes (main) go inactive, and memtable flushes dominate the remaining
  of the task_quota. Compactions are present in the active_queue, but
  the scheduler will pick memtable flushes (lower vruntime by a tiny
  margin)
- memtable flushes actually disrespect the task quota, running for
  ~450usec.

At this point, because memtable flushes have very small shares and it
ran for a long time its vruntime will be way more than what compactions
have. Note that this is exactly what we want: since memtable flushes had
run for a long time, we want to equalize it with compactions over time.

The problem is that after memtable flushes finish running,
_last_vruntime is set to that gigantic vruntime. At this point writes
(main group) become active and will inherit that vruntime (minus a task
quota).

Compactions are still *way* behind that vruntime and will dominate the
CPU for a long time, until those things equalize.

The fix for that is, in my view, surprisingly simple: instead of
accounting _last_vruntime as the vruntime the last task group had *after*
it ran, let _last_vruntime be the vruntime the task group had *before*
it ran.

With this change, we can at the same time reduce the dependency of the
adjustment algorithm to the task quota -- since even if a task group
disrespected it we will still be using a baseline of time before that
happened, and guaranteeing that we won't be too far off from the other
task groups that are active. After all, if we picked a task group to
execute that means that in that point of time it had the lowest
vruntime.

I have tested this change against a build of ScyllaDB patched to make
use of the CPU scheduler, running in a single CPU. The workload is a
write-only workload with a limit of 25000 writes/second with small
payloads being written by 20 threads.

In the unpatched version, we see severe drops when compactions run:

type,      total ops,    op/s,    pk/s,   row/s,    mean,     med,     .95,     .99,    .999
...
total,      12752234,   24999,   24999,   24999,     0.6,     0.6,     1.0,     1.2,     3.6
total,      12763744,   11466,   11466,   11466,     1.6,     0.8,     5.2,     5.5,     5.9
total,      12768059,    4291,    4291,    4291,     4.6,     5.1,     5.6,     6.3,     6.6
total,      12771931,    3864,    3864,    3864,     5.1,     5.5,     6.0,     6.5,     6.9
...

While in the patched version, those drops are totally gone.

We can see the difference more clearly in the execution summary:

Unpatched:

Results:
op rate                   : 22166 [WRITE:22166]
partition rate            : 22166 [WRITE:22166]
row rate                  : 22166 [WRITE:22166]
latency mean              : 0.6 [WRITE:0.6]
latency median            : 0.5 [WRITE:0.5]
latency 95th percentile   : 0.9 [WRITE:0.9]
latency 99th percentile   : 6.5 [WRITE:6.5]
latency 99.9th percentile : 11.2 [WRITE:11.2]
latency max               : 19.9 [WRITE:19.9]
Total partitions          : 13299383 [WRITE:13299383]
Total errors              : 0 [WRITE:0]
total gc count            : 0
total gc mb               : 0
total gc time (s)         : 0
avg gc time(ms)           : NaN
stdev gc time(ms)         : 0
Total operation time      : 00:09:59
END

Patched

Results:
op rate                   : 25011 [WRITE:25011]
partition rate            : 25011 [WRITE:25011]
row rate                  : 25011 [WRITE:25011]
latency mean              : 0.6 [WRITE:0.6]
latency median            : 0.5 [WRITE:0.5]
latency 95th percentile   : 0.8 [WRITE:0.8]
latency 99th percentile   : 1.0 [WRITE:1.0]
latency 99.9th percentile : 2.3 [WRITE:2.3]
latency max               : 23.1 [WRITE:23.1]
Total partitions          : 15006295 [WRITE:15006295]
Total errors              : 0 [WRITE:0]
total gc count            : 0
total gc mb               : 0
total gc time (s)         : 0
avg gc time(ms)           : NaN
stdev gc time(ms)         : 0
Total operation time      : 00:09:59
END

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180202191945.11075-1-glauber@scylladb.com>