json_element: stream range object
This is a specific implementation for steaming range object as
json array.

For example, a map can be return in the API as an array of key/value.

If a user of the framework needs to return such a map, this
implementation will save the need to translate the map to a vector first
and then stream it.

Instead, the helper function stream_range_as_array would capture the
object (in the example, a map) iterate over it and use a mapping
function to get the expected result.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <20180207142958.26140-1-amnon@scylladb.com>