net/stack: Don't destroy server_socket_impl with in-progress accept
accept()ing a connection captures a raw pointer to server_socket_impl,
so if the unique_ptr keeping it alive is destroyed before the accept
completes, we later get a use-after-free. This can happen if the
reactor is preempted while checking if the socket is readable, before
doing the actual accept.

Signed-off-by: Duarte Nunes <duarte@scylladb.com>