seastar-addr2line: improve UX for bactraces read from stdin
Previously when supplying backtraces on stdin the user had to press
Ctrl+D after entering the backtrace to trigger their resolution and then
kill the process with Ctrl+C. This is very bad user experience.
To solve this a new StdinBacktraceIterator is added which reads stdin
char-by-char and stops as soon as it discovers two line feeds in a row
or a Ctrl+D (empty string). This way if the user presses either an
additional Enter or a Ctrl+D after entering the backtrace it will
trigger the resolution and will stop the script.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <e21e95df0be88831990452c4356259d77b882c6f.1518594779.git.bdenes@scylladb.com>