io scheduler: continuation free dispatches, part 1
We issue batched I/O from pollers, and check for their readiness once
again in the same poll cycle. However, that process involves a couple of
continuations; making semaphores ready, chaining futures to indicate
ability to dispatch, etc.

It would be much preferrable if we could dispatch everything from the
pollers, so we don't have to go back to the task queue to make things
ready and then dispatch in the next cycle - this should reduce latency
of dipatches in a couple of situations and gain us some efficiency.

For shards dispatching remotely, the SMP queue continuations are still
in effect - I plan to handle that in a later date. But for shards
dispatching locally, we are now fully continuation free:

- the _io_context_available semaphore is gone, as capacity enforcement
  is now done totally through the I/O Scheduler
- the fair_queue semaphore is gone, replaced now by a notification
  mechanism where the reactor explicitly tell us about requests that
  finished so we can dispatch more.
- Instead of notifying about readiness through a future, so we could
  chain the request execution, the fair queue now executes a function
  directly. Also, promise created in submit_io is now created in
  queue_request() itself, so we don't have to chain two continuations
  per one I/O request.

To demonstrate the effects of this patch, I have run a test with 250usec task
quotas in a single shard.  The first test runs against a fast NVMe, with a load
the shouldn't bottleneck it. With constant parallelism and outside the
saturation point, the latency improvements deliver 10 % more performance for
the reads.

Before:
 Class 0(latency_reads: 100 shares, 4096-byte RAND READ, 40 concurrent requests, NO think time)
  Throughput         :   725795 KB/s
  Lat average        :      190 usec
  Lat quantile=  0.5 :      179 usec
  Lat quantile= 0.95 :      278 usec
  Lat quantile= 0.99 :      466 usec
  Lat quantile=0.999 :      670 usec
  Lat max            :     1827 usec

After:
 Class 0(latency_reads: 100 shares, 4096-byte RAND READ, 40 concurrent requests, NO think time)
  Throughput         :   803999 KB/s
  Lat average        :      156 usec
  Lat quantile=  0.5 :      138 usec
  Lat quantile= 0.95 :      266 usec
  Lat quantile= 0.99 :      391 usec
  Lat quantile=0.999 :      614 usec
  Lat max            :    10098 usec

-------------------------------------------------------------------------------------------------

Signed-off-by: Glauber Costa <glauber@scylladb.com>