io_tester: use the CPU scheduler
I/O tester is an I/O testing utility. But I/O requests don't exist on a
vaccuum and it is important to be able to simulate CPU issues that may
affect I/O latencies as well. As a first step, we'll change the meaning
of the parameter "shares" so that they also represent CPU shares - not
just I/O.

This will allow us to evaluate how is it that competing CPU users affect
latency profiles.

Signed-off-by: Glauber Costa <glauber@scylladb.com>