io_tester: add the ability to introduce CPU loads
The io-tester tool is a helpful tool to simulate I/O loads. However, it
is at the moment restricted to purely I/O loads. In practice, I/O
continuations have to fight for time for CPU continuations which will
some times behave well, and sometimes won't.

To be able to properly simulate those loads, we'll add a "cpu" request
type that is characterized by:
 - for how long it does some CPU computation (in here just a busy loop)
 - for how long it thinks before it fires another continuation

Signed-off-by: Glauber Costa <glauber@scylladb.com>