future: make future::wait() a supported function
future::wait() is marked as an internal function, but it can be
legitimately used in threads (similar to future::then_wrapped()
outside threads). This is useful if we want to check whether
the future failed before calling future::get() or future::get0().

To make it generally useful, remove the assumption that the
future is not yet available and check for it explicity. The
non-checking version is retained as a private do_wait(), for
use in future::get() (the only caller).

Example:

    future<> f1 = some_function();
    f1.wait();
    if (f1.failed()) {
        return -1;
    }
    return 0;
Message-Id: <20180305103536.27035-1-avi@scylladb.com>