tutorial: support, and start using, links to other sections
This patch adds supports for links to other sections in the tutorial.
These links look like [Lifetime management], i.e., the title of the
section (can the title of a chapter or of a smaller subsection) in
square brackets.

The links work on both pdf and HTML outputs (generated by the ninja
targets doc/tutorial.pdf and doc/tutorial.html, respectively).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180305143618.9808-1-nyh@scylladb.com>