tutorial: add build target for split HTML file
The single-HTML-page tutorial is getting long, and people have asked to
also provide a multi-page HTML, where each chapter is in a separate
HTML page.  This is what this patch provides: Building

    ninja doc/split

Now builds a directory doc/split/ with a separate HTML file per chapter.
The index.html file contains the table of contents, and each file <N>.html
contains chapter <N>. The table of contents points to the individual
chapters (and also to sections in them), and each chapter also points back
to the table of contents, as well as the previous and next chapters.
Links between different chapters are also fixed to correctly point to
the different documents.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180305144645.10378-1-nyh@scylladb.com>