io_tester: use fmt instead of c++ iostreams
Not only fmt is better, there is an immediate rationale for doing this
now: the code as is need a hack to print from class_data using
inheritance. I am about to turn class_data into an abstract class, so
this will become significantly more complex.

Better port this now.

Signed-off-by: Glauber Costa <glauber@scylladb.com>