rpc: Make connection class completely independent of rpc::protocol template
Currently connection class holds a reference to rpc::protocol instance
and thus has to be a template since rpc::protocol is one. The only
reason the reference is needed is to get to a serializer instance which
is held in rpc::protocol. We can break the dependency by holding a type
erased pointer to serializer in the connection class and converts it to
correct type at the place of use since every place where connection is used
already knows serializer type.