rpc: move server class out of protocol
Since connection class the server class inherits from is not part of
rpc::protocol template any more the server class can also be moved
out of it (and a lot of code moves into .cc file with it!). Only shim
is left for compatibility with existing user code since it relies on
rpc::protocol::server class existence.