memory: avoid unconditional calls to __tls_init
Memory allocator uses thread-local objects that are not trivially
constructible. This means that before each access it has to be ensured
that they have been already properly initialised.

GCC does that by injecting uncodnitional calls to __tls_init. This is
particularly bad for applications that are heavy on small allocations.

This patch alleviates the problem by checking if the thread-local
object has already been initialised and calling __tls_init only if it
hasn't.

Message-Id: <20180308142907.16208-1-pdziepak@scylladb.com>