routes: Add alias support
Sometimes it is required to change an API (for example when fixing a
typo in the path).

Typically, this will involve adding a new API with the correct spelling
deprecate the old API and support both for sometimes.

This patch adds an aliasing option to the routes object, so supporting both
APIs would only require setting one of the APIs as an alias.

Example:

if new_api and old_api are both path_description (auto-generated by the
json2code) and r is the route object:

if old_api was used to regsiter a handler.

Now it is possible to add an alias for that path by calling:

r.add_alias(old_api, new_api);

Signed-off-by: Amnon Heiman <amnon@scylladb.com>