cpu scheduler: don't penalize first group to run
We are keeping a variable we use to calculate the time for how long a
group run, t_run_completed, outside run_some_tasks.

When we go idle, we update that variable right before we go back to
run_some_more_tasks. That is correct, and presents no issues.

But when we are not idle, we'll run pollers and go back to
run_some_tasks right after. t_run_completed will still have the last
time assigned to it the last time it ran, and when we compute, for the
first group:

   t_run_started = t_run_completed
   ... run group ...
   vruntime = to_vruntime(now() - t_run_started)

we will end up in a situation where t_run_started is t_run_completed
before pollers ran. This means we will assign a vruntime to the tasks
much higher than what it actually spent.

Although keeping that as a variable save us one call to now(), it isn't
really accurate and we should lose it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180314000246.32294-1-glauber@scylladb.com>