reactor: fix sleep mode
Avi recently noticed that the I/O rework patches broke poll mode. The
reason is that I got confused by the comment on try_enter_interrupt_mode
for the I/O pollers and tested for the *presence* of in flight AIO
requests as a condition to enter, instead of their absence.

This patch fixes that and also make the comment more clear, in favor of
future readers.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180314121110.6397-1-glauber@scylladb.com>