reactor: pass naked pointer to submit_io
We had previously explicitly coded submit_io so that it takes ownership
of the unique_ptr containing the promise and releases it. While that is
good, it is making exeception handling a lot more complex: if the
exception happens inside submit_io, we can't set the exception promise
because the unique_ptr would have been destroyed.

It is not as simple as moving the try/catch block inside submit_io
either, since the exception can also happen before that, and furthermore
we have to make adjustments to the queue.

The least worst solution seems to me to compromise on the principle that
led us to move the unique_ptr and handle all exceptions in the caller,
which will be responsible for releasing and destroying the unique_ptr
whatever happens.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180320140445.13309-1-glauber@scylladb.com>