reactor: pass a descriptor pointer to submit_io
We are currently passing a pointer to promise<io_event> to submit_io.  I
would like to enhance the information that is passed down to include a
bit more information about the I/O: whether or not it is a read or a
write, its size, etc.

For that reason, I am wrapping that promise in a descriptor and changing
the signature of submit_io to take a pointer to that descriptor.

Signed-off-by: Glauber Costa <glauber@scylladb.com>