io_queues: move capacity planning out of resource manager
Currently we defer to the code in resource.cc to distribute the I/O
queues into the existing shards, and while we are on it, also figure
out the capacity for each.

We are about to make the capacity decision a bit more complex, so it
would be beneficial to separate those things. We can make the code in
resource.cc do only what it does best, which is distribute things within
the available resources.

Capacity planning can be done in reactor.cc

Signed-off-by: Glauber Costa <glauber@scylladb.com>