http routes: Add parameters to path when adding alias
When adding an alias to the routes we use the same parsing that we use
when parsing an incoming request.

This assume that path parameters are part of the path, but the code
generator removes those parameters when it creates the path_description
path parameter.

To solve that, we need to add a path parameter for each parameter.

For example, assume that the original definition was:

/my/path/{key1}/{key2}

The code generator will create a path:
/my/path
with two path parameters (key1, key2).

To resolve the handler when adding an alias we will use the following
path:
/my/path/k/k

Note that the actual value of the key is meaningless, and there is no
performance aspect to it, as this is done once during startup.

Fixes #428

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <20180408115326.28039-1-amnon@scylladb.com>