rpc: introduce streaming over rpc
This patch introduce streaming capability to rpc. Each stream has
a sink and a source both on a client and on a server. To send
data one does (assuming the code runs in seatar::async context):
      while (has_data()) {
          int data1 = get_data1();
          long data2 = get_data2();
          ...
          sink(data1, data2, ...).get(); // sends data
      }
      sink.close().get(); // closes stream

 to receive
      while (true) {
          stdx:optional<std::tuple<int, long, ...>> data = source().get();
          if (!data) {
             // unengaged optional means EOS
             break;
          } else {
             // process data
          }
      }

How client and server obtain correspondent sink and source?
It can be best explained by an example. Suppose we want to create a
stream of ints in one direction and strings in another.

Server code should register rpc handler like this:

    // aux_data is just an auxiliary data that can be used to describe
    // what the stream is about, it can be of any type and can also
    // include more than one parameter, its placement relative to
    // source parameter is irrelevant too.
    rpc_proto.register_handler(1, [] (int aux_data, rpc::source<int> source) {
        rpc::sink<sstring> sink = source.make_sink<sstring>();
        // use sink and source asynchronously
        return sink;
    });

Client code:
   auto rpc_call = rpc_proto.make_client<rpc::source<sstring> (int, rpc::sink<int>)>(1);
   auto f = rpc_client.make_stream_sink<int>().then([rpc_call] (rpc::sink<int>&& sink) {
       return rpc_call(rpc_client, aux_data, sink).then([sink = std::move(sink)] (rpc::source<sstring>&& source) mutable {
         // use sink and source
       });
   });

Note that sink and source are not movable between shards and all sending
and receiving is done by a shard where corresponding sink and source
were obtained.

If rpc client used to create streams terminates (because of an error or
explicitly closed) all associated streams will terminate too.

To enable streaming capability one should provide unique
streaming_domain (uint64_t value) in server_options when creating
an rpc server. Servers on different shards that constitute a single
logical server must belong to the same streaming domain.