fair_queue: encapsulate properties in a request descriptor
We current allow the caller of the fair_queue to specify a weight
for the request - that will be used to equalize the throughput that
each class gets.

Soon we are about to inform the fair_queue about a lot more than just
the weight. So wrap that information in a descriptor.

Signed-off-by: Glauber Costa <glauber@scylladb.com>