reactor: extend the fair queue with size and iops limits
Requests are not all equal and can be of different sizes and have
different weights for the purpose of counting IOPS (reads for instance
tend to be cheaper than writes).

Once we track those requests at creation time and add them to the I/O
descriptor, we can then notify the fair queue of their completion with
full information. This will be used to implement per-type limits as well
of separate throughput/bandwidth limits.

After this patch, the fair queue effective acquires the capacity of being
limited on both throughput and iops. The old, simple capacity is kept both for
compatibility and because the upper layer expects us to limit the maximum
amount of requests in flight at any given time so as not to overflow the
dispatch structures.

Signed-off-by: Glauber Costa <glauber@scylladb.com>