reactor: gracefully exit if exception happens during initialization
If an option fails to be parsed, we exit seastar gracefully, courtesy of
boost::program_options. However, if an option is syntactically valid but
found to be invalid during runtime, then things are not that graceful.

Let's see the example of passing a number of smp (but still a valid
integer) that is found by the resource parsing code to be unreasonably
big. This is what happens:

$ ./build/release/apps/seastar_app  --smp 45
terminate called after throwing an instance of 'std::runtime_error'
  what():  insufficient processing units
Aborting.
Backtrace:
  0x00000000004cabfc
  0x00000000004caea5
  0x00000000004caf53
  /lib64/libpthread.so.0+0x0000000000012aef
  /lib64/libc.so.6+0x000000000003766a
  /lib64/libc.so.6+0x0000000000039380
  /lib64/libstdc++.so.6+0x0000000000092024
  /lib64/libstdc++.so.6+0x000000000008fc15
  /lib64/libstdc++.so.6+0x000000000008fc60
  /lib64/libstdc++.so.6+0x000000000008fea3
  0x00000000006272b9
  0x000000000058fb3f
  0x000000000066f688
  0x0000000000671f9b
  0x0000000000418117
  /lib64/libc.so.6+0x0000000000021009
  0x00000000004193e9
Aborted (core dumped)

This includes the exception message (good), but also includes a
backtrace that is not at all helpful for the case of simply seeing an
invalid value and worse of all - a coredump. We should catch those cases
explicitly and gracefully terminate.

After this patch, I see the following output for that same command line:

$ ./build/release/apps/seastar_app  --smp 45
Could not initialize seastar: insufficient processing units

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180417223823.22571-1-glauber@scylladb.com>