io_queue: determine concurrency based on the disk's visible throughput
Problems with the Seastar I/O Scheduler
=======================================

The I/O Scheduler has been a cornerstone of Seastar applications since
its inception, in particular ScyllaDB, with its first commit dating back
to January 2016. Two years of production experience with the I/O
Scheduler have solidified its importance but also exposed some of its
shortcomings - which I plan to address in this series.

Mainly, those shortcomings are:

1) It is very hard for users to reason about what --max-io-requests
mean. Very often people get confused believing that higher is always
better and keep being puzzled that their new, faster NVMe has a lower
--max-io-requests than their older, slower SSDs.

Also, when things get mistuned - which is just a fact of life in the
real world, it is very hard for people to reason about what to do with
that parameter. Should I lower or raise it? By how much? What should I
expect?

2) It is hard to tune. In many situations, iotune simply can't get
--max-io-requests right. It's a noisy statistical process looking for an
indirect quantity. Users complain that the results are sometimes wildly
different among very similar disks.

3) It doesn't differentiate between reads and writes. Many disks aren't
able to serve both reads and writes at equal speeds. There is a single
parameter for the disk (--max-io-requests) and it is measured for reads.
We have seen disks with ratios higher than 4:1 read:write speed, meaning
that for write-heavy workloads the disk gets severy overloaded.

4) It doesn't take request sizes into account. Our single parameter
--max-io-requests is usually discovered based on 4kB reads. But larger
request take longer and have a much lower sweet spot for parallelism,
meaning that at the tuned parallelism the disk also gets severely
overloaded and a latency request coming right after will face long wait
times.

A Key observation
=================

There is a little detail that we can note that make our lifes a little
easier: Little's Law for closed systems. It states that:

  X = E[t] * T,

where X is the concurrency, E[t] is the expected time to complete a
request (latency), and T is the throughput. This is a key observation,
for the following reasons:

1) If we know what is the maximum throughput of the device, then we can
know that increasing concurrency can only increase latency. Conversely,
reducing concurrency will reduce latency, provided that this reduction
does not result in a decrease in throughput.

2) In steady state, we can only do I/O at each task_quota_ms. What that
means is that even if our requests are as fast as a couple of
microseconds, we very likely won't be able to consume them after at
least one task quota.

Assuming for example a task quota of 0.5ms, and a single CPU driving
1M IOPS, we can write:

  X = 0.5ms * 1M IOPS = 5

In other words, if we allow 5 requests to be sent to the disk
concurrently, they are expected to take 0.5ms to complete. Even if the
individual request is much faster than that we are stil ok - we wouldn't
be able to consume it sooner. In practice, we can include a factor
(current 1.5) in the calculation, since not all requests will be
finished by the task quota (Little Law's deals with average). We want to
keep busy at all times, so we will slightly overshoot.

From the get go, this approach addresses the first two problems listed
in the introduction of this article:

1) Users can very easily reason about throughput. If things get mistuned
or the disks deteriorate, it is very easy to, by observation, find out
what the real throughput is and adjust. Users can very easily reason
about what does it mean to increase the throughput by a certain margin
or conversely decreasing it.

2) Throughput is a direct measurement and there is no complex process
involved in figuring out what that is.

But problems 3 and 4 can also be addressed by noting that Little's Law
is valid for the throughput of any process: we can state read
throughput and write throughput, or state them in MB/s, and just limit
the amount of requests sent to disk based on which one bottlenecks
first.

For example, my local SSDs can do 95k IOPS and 512MB/s. 95k/s 4kB
requests only amount to 371 MB/s. If all requests are 4kB in size we
will hit the IOPS bottleneck but not the throughput bottleneck. But if
the requests are larger, 128kB in size, we will hit the bandwidth
bottleneck while doing 4k IOPS.

While the current I/O Scheduler would keep 32 requests in parallel
for all of those situations, the proposed version would (assuming a
0.5ms task quota) keep

  x = 95 kIOPS * 1.5 * 0.5 ms = 70 concurrent 4kB requests, and

  y = 512 MB * 1.5 * 0.5 ms = 402kB, or ~3 concurrent 128kB requests.

Note how in the first scenario, latency is likely to be higher than what
the current I/O Scheduler produces - but we are assuming we won't
consume before the next task quota anyway. But in the second scenario,
latencies will be much lower than what the current I/O Scheduler
produces.

Results
=======

To demonstrate this, I have run a highly parallel write workload with
128kB requests against a latency-bound small read workload - with some
CPU load in the background. As a single CPU is able to drive this disk
to its maximum, I will show results from a single CPU since that is
easier to see.

We can see that this proposal is able to keep throughput at the same max
level as before, but decrease latencies in every percentile, with the
averages going down almost 8-fold.

Current seastar upstream
=====================================================================================================

Class  0 (big_write: 10 shares, 131072-byte SEQ WRITE, 100 concurrent requests, NO think time)
  Throughput         :   518684 KB/s
  Lat average        :    24662 usec
  Lat quantile=  0.5 :    24640 usec
  Lat quantile= 0.95 :    24719 usec
  Lat quantile= 0.99 :    24879 usec
  Lat quantile=0.999 :    36128 usec
  Lat max            :    43170 usec

Class  1 (latency_reads: 100 shares, 4096-byte RAND READ, 1 concurrent requests, 10000 us think time)
  Throughput         :  219.994 KB/s
  Lat average        :     8143 usec
  Lat quantile=  0.5 :     7880 usec
  Lat quantile= 0.95 :     9085 usec
  Lat quantile= 0.99 :    12509 usec
  Lat quantile=0.999 :    15134 usec
  Lat max            :    24171 usec

Class  2 (cpu_hog: 10 shares, 90 us CPU execution time, 10 concurrent requests, 10 us think time)
  Throughput         :  10656.3 continuations/s

Proposed
=====================================================================================================

Class  0 (big_write: 10 shares, 131072-byte SEQ WRITE, 100 concurrent requests, NO think time)
  Throughput         :   514262 KB/s
  Lat average        :    24873 usec
  Lat quantile=  0.5 :    24830 usec
  Lat quantile= 0.95 :    24930 usec
  Lat quantile= 0.99 :    25249 usec
  Lat quantile=0.999 :    36347 usec
  Lat max            :    41952 usec

Class  1 (latency_reads: 100 shares, 4096-byte RAND READ, 1 concurrent requests, 10000 us think time)
  Throughput         :  347.414 KB/s
  Lat average        :     1452 usec
  Lat quantile=  0.5 :     1324 usec
  Lat quantile= 0.95 :     2334 usec
  Lat quantile= 0.99 :     5936 usec
  Lat quantile=0.999 :     7584 usec
  Lat max            :    11026 usec

Class  2 (cpu_hog: 10 shares, 90 us CPU execution time, 10 concurrent requests, 10 us think time)
  Throughput         :  10607.8 continuations/s

Signed-off-by: Glauber Costa <glauber@scylladb.com>