cmake: Do not enable warnings transitively
In CMake, target-specific properties can be specified as either
`PUBLIC`, `PRIVATE`, or `INTERFACE`.

`PUBLIC` properties of a library apply transitively to targets which
link against the library.

In general, compilation options which are more restrictive (like
enabling warnings) should not be transitive. This is because we should
not cause application code (unrelated to Seastar) to fail to compile
just because it links against Seastar.

Conversely, compilation options which loosen restrictions (like
disabling warnings) should be transitive. This is because an application
that enforces restrictions for its code should not be prevented from
compiling because of failures in Seastar to adhere to those
restrictions. This of course implies that the number of loosened
restrictions in Seastar should be minimized to the greatest extent
possible.

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <f6dc5d18d9d9bff7a59d2cbbdb3f0d963384e273.1524842563.git.jhaberku@scylladb.com>