iotune: rewrite iotune and generate property files according to new I/O scheduler format
This patch changes iotune so that it can work with our new way of
specifying disks. It is presented here as a separate file to make
reviewer's lives easier.  A follow up patch will git-mv it to iotune.cc

I am introducing here a dependency on the C++17 std::filesystem
(experimental in gcc), mainly because the interface is way too nice not
to use - specially for the range directory scans.

The basic idea here is that we will using fully sequential I/O figure
out the bandwidth limits of the disk, which can be done in a single CPU.
And then with random I/O from all CPUs we will figure out the IOPS
limits of the devices.

Because we are still, at the moment, relying on I/O queues (they should
soon go away when we dispatch from all shards, but not that soon), I am
keeping that in an old-school options file in the same format as what
current iotune generates.

Aside from the new method for calculating the properties, I am also
doing discovery of some of the disks properties from sysfs to learn
how to best interact with the disk.

Signed-off-by: Glauber Costa <glauber@scylladb.com>