Tutorial: in PDF version of tutorial, better backtick typesetting
The PDF version of the tutorial, typeset using TeX, shows backticks (`)
by default as curved quote characters. In code samples, those are
ugly and confusing for readers. This patch uses the "upquote" style to
fix that (just in verbatim environments, namely code samples).

After this patch, a user who wants to build a PDF tutorial (ninja
doc/tutorial.pdf) needs the "upquote" TeX package, which should be
easy to install on modern Linux distributions - e.g., in Fedora, just
do "yum install texlive-upquote". I could have made the use of this
package optional, but I don't want to: I think without it the resulting
PDF looks bad, and the user should get an error if this package is
missing and go and install it.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180429084434.13356-1-nyh@scylladb.com>