reactor: add rudimentary task histogram reporting
If enabled (by a macro), report a histogram of task types executed
every 1 million tasks. This gives a developer-level view of what is
going on.

We can probably do a runtime version of this with statically defined
tracepoints and a fancy script, but it will likely perturb
performance far too much to be useful.
Message-Id: <20180430095630.25348-1-avi@scylladb.com>