future: remove allocation from future::get() thread context switch
When we call future::get() in a thread, and the future is not available,
we put the thread to sleep and attach a continuation to the promise. The
continuation will move the result to the future, and wake up the thread.

We can optimize this by allocating the continuation on the thread's
stack instead of on the heap, since it won't be needing it while
asleep.

Reduces thread_context_switch from ~128ns to ~106ns on my machine.

Tests: unit (release)
Message-Id: <20180430115930.16225-1-avi@scylladb.com>